Refactor calculation archive management
This way is actually a lot more robust and less confusing.  Make
archives only if the calculation finished and ran into a fixable VASP
error or didn't reach required accuracy.  The archive creation logic is
pulled into check_calc, and as a benefit, CalcManagers' setup_calc and
VaspInputCreator's make_archive_and repopulate need less code. The
VaspInputCreator now looks to see if an archive exists, and if it does,
it pulls the source structure (CONTCAR) from the most recent archive.