Fix ncore for multiprocessing
Default to 4 processes. Setting the default to the number of cores
can be problematic when running on a supercomputer, where there
are lots of cores but many of them may not be available for
VaspManager. Add a warning to tell users to set ncore themselves
instead of letting VaspManager run into troublesome threading errors.