Update number of significant figures for POSCAR writing
In very, very rare cases, it turns out that rounding the significant
figures to 8 instead of 16 will break the symmetry of the cell. I've
only hit this issue for certain R-3c cells, so I imagine this has
something to do with their irrational symmetry operations and floating
point precision.